improve build process

+++ db/migrations/20141221205038_add_account_to_user.js
@@ -1,13 +1,19 @@
 'use strict';
 
 exports.up = function(knex, Promise) {
-    return knex.schema.table('users', function (table) {
-        table.integer('account').defaultTo(0);
-    })
+    knex.schema.hasTable('users').then(function(exists) {
+        return knex.schema.table('users', function (table) {
+            table.integer('account').defaultTo(0);
+        })
+    });
 };
 
 exports.down = function(knex, Promise) {
-    return knex.schema.table('users', function (table) {
-        table.dropColumn('account');
-    })
+    knex.schema.hasColumn('users',"account").then(function(exists) {
+        if(exists){
+            return knex.schema.table('users', function (table) {
+                table.dropColumn('account');
+            });
+        }
+    });
 };
